backend/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
youtube-transcript-api==1.2.2
google-generativeai==0.8.3
numpy==2.3.3
python-dotenv==1.0.1
pydantic==2.10.4
yt-dlp==2024.12.23
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
